requirements.txt
python-docx
Pillow
pdfplumber